Update copyright statement
Added back the asterisks since it looks a bit nicer this way. The copyright statement was previously updated to a format in which the initial spaces would be preserved when the file was re-indented in Xcode, instead of un-indenting as the previous version did. Turns out that what preserves the space is the inclusion of the extra initial asterisk, i.e. "/**" instead of "/*" so added back the asterisks to the intermediate lines.