Preferred, Not Preferred Ordering Maintained
When example are given throughout the guide, it shows the preferred instance before the not preferred instance. However, in [unused code](https://github.com/raywenderlich/swift-style-guide#unused-code) this pattern is reversed, which may lead to confusion. For the sake of clarity and consistency, the Preferred section is now placed above the Not Preferred section, as is done throughout the rest of the style guide.